hotels資料表匯入完成

+++ staytion/src/main/java/com/example/demo/entity/Hotel.java
@@ -12,13 +12,13 @@ import jakarta.persistence.Table;
 public class Hotel {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
 	@Column(name = "hname")
 	private String hname;
 
 	@Column(name = "district_id")
-	private Integer districtId;
+	private Long districtId;
 
 	@Column(name = "address")
 	private String address;
@@ -38,11 +38,11 @@ public class Hotel {
 	@Column(name = "longitude")
 	private Double longitude;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -54,11 +54,11 @@ public class Hotel {
 		this.hname = hname;
 	}
 
-	public Integer getDistrictId() {
+	public Long getDistrictId() {
 		return districtId;
 	}
 
-	public void setDistrictId(Integer districtId) {
+	public void setDistrictId(Long districtId) {
 		this.districtId = districtId;
 	}
